back-end/pyproject.toml
[tool.poetry]
name = "back-end"
version = "0.1.0"
description = ""
authors = ["Carlos Daniel Barros <barros.carlosds@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
django-filter = "^2.3.0"
djangorestframework = "^3.11.1"
python-decouple = "^3.3"
djangorestframework-jwt = "^1.11.0"
dj_database_url = "^0.5.0"
django-cors-headers = "^3.5.0"
psycopg2 = "^2.8.6"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"
pylint = "^2.6.0"
pylint-django = "^2.3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
